LatticeBoltzmannModel

Documentation for LatticeBoltzmannModel.

LatticeBoltzmannModel.AbstractLatticeBoltzmannConfiguration

LatticeBoltzmannModel.D2Q9

LatticeBoltzmannModel.LatticeBoltzmann

LatticeBoltzmannModel.Point

LatticeBoltzmannModel.curl

LatticeBoltzmannModel.example_d2q9

LatticeBoltzmannModel.get_equilibrium_cell

LatticeBoltzmannModel.get_momentum

LatticeBoltzmannModel.step!

LatticeBoltzmannModel.AbstractLatticeBoltzmannConfiguration — Type

AbstractLatticeBoltzmannConfiguration{D, N}

An abstract type representing a lattice Boltzmann configuration in D dimensions with N velocities.

LatticeBoltzmannModel.D2Q9 — Type

D2Q9 <: AbstractLatticeBoltzmannConfiguration{2, 9}

D2Q9 lattice Boltzmann model, which is defined in 2-dimensional space, and has 9 velocities.

LatticeBoltzmannModel.LatticeBoltzmann — Type

LatticeBoltzmann{D, N, T, C, M, B}

A lattice Boltzmann simulation with D dimensions, N velocities, and lattice configuration CFG.

Fields

config::C: lattice Boltzmann configuration

grid_cur::M: current grid

grid_old::M: old grid

barrier::B: barrier matrix

LatticeBoltzmannModel.Point — Type

Point{D, T}

A point in `D`-dimensional space with coordinates of type `T`.

LatticeBoltzmannModel.curl — Method

curl(u::AbstractMatrix{Point2D{T}})

Compute the curl of the momentum field in 2D.

LatticeBoltzmannModel.example_d2q9 — Method

example_d2q9(; height = 80, width = 200, u0 = Point(0.0, 0.1))

A D2Q9 lattice Boltzmann simulation example. A simple linear barrier is added to the lattice.

Arguments

height::Int: height of the lattice

width::Int: width of the lattice

u0::Point2D: initial and in-flow speed

LatticeBoltzmannModel.get_equilibrium_cell — Method

get_equilibrium_cell(config::AbstractLatticeBoltzmannConfiguration, rho, u)

Compute the equilibrium density of the fluid from the total density and momentum.

LatticeBoltzmannModel.get_momentum — Method

get_momentum(config::AbstractLBConfig, c::Cell)

Compute the momentum of the fluid from the density of the fluid.

LatticeBoltzmannModel.step! — Method

step!(lb::LatticeBoltzmann)

Perform a single step of the lattice Boltzmann simulation.