LatticeBoltzmannModel

Documentation for LatticeBoltzmannModel.

LatticeBoltzmannModel.AbstractLatticeBoltzmannConfiguration

LatticeBoltzmannModel.D2Q9

LatticeBoltzmannModel.Point

LatticeBoltzmannModel.get_momentum

LatticeBoltzmannModel.AbstractLatticeBoltzmannConfiguration — Type

AbstractLatticeBoltzmannConfiguration{D, N}

An abstract type representing a lattice Boltzmann configuration in D dimensions with N velocities.

LatticeBoltzmannModel.D2Q9 — Type

D2Q9 <: AbstractLatticeBoltzmannConfiguration{2, 9}

D2Q9 lattice Boltzmann model, which is defined in 2-dimensional space, and has 9 velocities.

LatticeBoltzmannModel.Point — Type

Point{D, T}

A point in `D`-dimensional space with coordinates of type `T`.

LatticeBoltzmannModel.get_momentum — Method

get_momentum(lb::AbstractLBConfig, c::Cell)

Compute the momentum of the fluid from the density of the fluid.